Fix: unit tests broken after refactoring
