+++ Makefile
@@ -26,7 +26,7 @@ test:
 	export COMPOSE_PROJECT_NAME=server-tests && \
 	$(TEST_COMPOSE_CMD) build && \
 	$(TEST_COMPOSE_CMD) run --rm server-tests \
-		python -m pytest --import-mode=importlib -v -s ${TEST_ARGS} ; \
+		python -m pytest --import-mode=importlib -v -s ${TEST_ARGS} && \
 	$(TEST_COMPOSE_CMD) stop
 
 test-clean:
